# contracts/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API behavior, compatibility rules, and endpoint contract requirements.",
    "version": "1.25.1"
  },
  "paths": {
...
    "/api/spool/{namespace}/{query_id}.parquet": {
      "get": {
        "summary": "GET /api/spool/{namespace}/{query_id}.parquet",
        "responses": {
          "200": {
            "description": "Contract response: application/octet-stream (parquet)"
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "410": {
            "description": "Error response (see contract error format)"
          }
        },
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-response-contract": "application/octet-stream (parquet)",
        "x-cdd-errors": "400/410"
      }
    },
...
    "/api/hold-history/config": {
      "get": {
        "summary": "GET /api/hold-history/config",
        "responses": {
          "200": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-errors": "500"
      }
    },
    "/api/hold-history/query": {
      "post": {
        "summary": "POST /api/hold-history/query",
        "responses": {
          "201": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "202": {
            "description": "Error response (see contract error format)"
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "410": {
            "description": "Error response (see contract error format)"
          },
          "500": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Contract request: JSON body (schema not machine-resolved; see contract)",
          "content": {},
          "x-cdd-unresolved": true
        },
        "x-cdd-errors": "202/400/410/500"
      }
    },
    "/api/hold-history/today-snapshot": {
      "post": {
        "summary": "POST /api/hold-history/today-snapshot",
        "responses": {
          "201": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "503": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Contract request: JSON body (schema not machine-resolved; see contract)",
          "content": {},
          "x-cdd-unresolved": true
        },
        "x-cdd-errors": "400/503"
      }
    },
    "/api/hold-history/view": {
      "get": {
        "summary": "GET /api/hold-history/view",
        "responses": {
          "200": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "410": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-errors": "400/410"
      }
    },
...
    "/api/resource/history/page": {
      "get": {
        "summary": "GET /api/resource/history/page",
        "responses": {
          "200": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-errors": "—"
      }
    },
    "/api/resource/history/export": {
      "get": {
        "summary": "GET /api/resource/history/export",
        "responses": {
          "200": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "410": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-errors": "400/410"
      },
      "post": {
        "summary": "POST /api/resource/history/export",
        "responses": {
          "201": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "410": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Contract request: JSON body (schema not machine-resolved; see contract)",
          "content": {},
          "x-cdd-unresolved": true
        },
        "x-cdd-errors": "400/410"
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "GenericSuccessResponse": {
        "type": [
          "object",
          "null"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {},
          "error": {},
          "meta": {
            "type": "object",
            "properties": {
              "timestamp": {
                "type": "string"
              },
              "app_version": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  }
}